+++ packages/vuikit/src/core/v-position/constants.js
@@ -1,0 +1,2 @@
+export const BEFORE_POSITION = 'v-vk-position:before'
+export const AFTER_POSITION = 'v-vk-position:after'