+++ packages/vuikit/src/core/v-position/events.js
@@ -1,4 +1,0 @@
-// before the element is positioned
-export const positionBefore = 'v-vk-position:before'
-// after the element is positioned
-export const positionAfter = 'v-vk-position:after'